Extract profile dropdown menu into profileDropdownMenu.tsx
